Fix MAC signature path
